if no file then no data

+++ dags/stellar_etl_airflow/build_load_task.py
@@ -96,11 +96,20 @@ def upload_to_gcs(data_type, prev_task_id, **kwargs):
     bucket_name = Variable.get('gcs_exported_data_bucket_name')
 
     logging.info(f'Attempting to upload local file at {local_filepath} to Google Cloud Storage path {gcs_filepath} in bucket {bucket_name}')
-    success = attempt_upload(local_filepath, gcs_filepath, bucket_name)
+    # TODO: consider adding a sanity check to make sure the filename shouldn't exist (ie no data to upload)
+    if os.path.exists(local_filepath):
+        success = attempt_upload(local_filepath, gcs_filepath, bucket_name)
+        fileExists = True
+    else:
+        logging.info('File does no exist, no data to upload')
+        success = True
+        fileExists = False
+    
     if success:
         #TODO: consider adding backups or integrity checking before uploading/deleting
-        logging.info(f'Upload successful, removing file at {local_filepath}')
-        os.remove(local_filepath)
+        if fileExists:
+            logging.info(f'Upload successful, removing file at {local_filepath}')
+            os.remove(local_filepath)
     else: 
         raise AirflowException('Upload was not successful')
